airule_agent/pyproject.toml
[project]
name = "airule-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "confluent-kafka>=2.10.1",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.41",
    "pydantic>=2.11.7",
    "redis>=6.2.0",
    "python-dotenv>=1.0.0",
    "asyncpg>=0.30.0"
]
